PyJEM scripts for lens and tilt velocity control
Python scripts based on the PyJEM package to freely control the CL2 and CL3 lenses values and choose the tilt velocity of your TEM stage. These scripts are provided as graphical user interfaces and they only work on JEOL microscopes that have the PyJEM package installed in the PC where the scripts are executed.